HistogramBinnings

Documentation for HistogramBinnings.

HistogramBinnings.LinEdgeVector

HistogramBinnings.LogEdgeVector

StatsAPI.fit

StatsBase.midpoints

HistogramBinnings.LinEdgeVector — Type

LinEdgeVector{T}(; lo = 1, hi = 10, nbins::Integer) where T <: Real

Construct a linearly spaced edge vector of type T. The edges are spaced evenly in linear space. lo and hi are the lower and upper limits of the histogram. nbins is the number of bins.

Examples

using StatsBase, HistogramBinnings

h = Histogram(LinEdgeVector(lo = 1, hi = 1_000_000, nbins = 60))
append!(h, vs)

HistogramBinnings.LogEdgeVector — Type

LogEdgeVector{T}(; lo = 1, hi = 10, nbins::Integer) where T <: Real

Construct a logarithmically spaced edge vector of type T. The edges are spaced evenly in log space. lo and hi are the lower and upper limits of the histogram. nbins is the number of bins.

Examples

using StatsBase, HistogramBinnings

h = Histogram(LogEdgeVector(lo = 1, hi = 1_000_000, nbins = 60))
append!(h, vs)

StatsAPI.fit — Method

fit(::Type{Histogram}, ::Type{T}, vs::AbstractVector;
    nbins = sturges(length(vs)),
    lo = minimum(vs),
    hi = maximum(vs)
) where T <: AbstractEdgeVector

Fit a histogram with edges T to the data vs.

T is either LogEdgeVector or LinEdgeVector. The edges are Int if the vs are Int otherwise they are Float64. lo and hi are the lower and upper limits of the histogram. nbins is the number of bins.

Examples

using StatsBase, HistogramBinnings
using Distributions

vs = floor.(Int, rand(Pareto(), 10000000))

h = fit(Histogram, LogEdgeVector, vs)

# or 

h = Histogram(LogEdgeVector(lo = 1, hi = 1_000_000, nbins = 60))
append!(h1, vs)

StatsBase.midpoints — Method

midpoints(r::AbstractEdgeVector{T}) where T

Return the midpoints of the bins in r.